+++ lib/customize/typography.php
@@ -18,9 +18,8 @@ add_action( 'init', 'mai_typography_customizer_settings' );
  * @return void
  */
 function mai_typography_customizer_settings() {
-	$handle        = mai_get_handle();
-	$section       = $handle . '-typography';
-	$global_styles = mai_get_global_styles();
+	$handle  = mai_get_handle();
+	$section = $handle . '-typography';
 
 	\Kirki::add_section(
 		$section,
@@ -30,6 +29,9 @@ function mai_typography_customizer_settings() {
 		]
 	);
 
+	$body_font_family = mai_get_default_font_family( 'body' );
+	$body_font_weight = mai_get_default_font_weight( 'body' );
+
 	\Kirki::add_field(
 		$handle,
 		[
@@ -37,10 +39,10 @@ function mai_typography_customizer_settings() {
 			'settings'    => 'body-typography',
 			'section'     => $section,
 			'label'       => __( 'Body', 'mai-engine' ),
-			'description' => __( 'Default: ', 'mai-engine' ) . mai_get_default_font_family( 'body' ) . ' ' . mai_get_default_font_weight( 'body' ),
+			'description' => __( 'Default: ', 'mai-engine' ) . $body_font_family . ' ' . $body_font_weight,
 			'default'     => [
-				'font-family' => mai_get_default_font_family( 'body' ),
-				'font-weight' => mai_get_default_font_weight( 'body' ),
+				'font-family' => $body_font_family,
+				'font-weight' => $body_font_weight,
 			],
 			'output'      => [
 				[
@@ -59,6 +61,9 @@ function mai_typography_customizer_settings() {
 		]
 	);
 
+	$heading_font_family = mai_get_default_font_family( 'heading' );
+	$heading_font_weight = mai_get_default_font_weight( 'heading' );
+
 	\Kirki::add_field(
 		$handle,
 		[
@@ -66,10 +71,10 @@ function mai_typography_customizer_settings() {
 			'settings'    => 'heading-typography',
 			'section'     => $section,
 			'label'       => __( 'Heading', 'mai-engine' ),
-			'description' => __( 'Default: ', 'mai-engine' ) . mai_get_default_font_family( 'heading' ) . ' ' . mai_get_default_font_weight( 'heading' ),
+			'description' => __( 'Default: ', 'mai-engine' ) . $heading_font_family . ' ' . $heading_font_weight,
 			'default'     => [
-				'font-family' => mai_get_default_font_family( 'heading' ),
-				'font-weight' => mai_get_default_font_weight( 'heading' ),
+				'font-family' => $heading_font_family,
+				'font-weight' => $heading_font_weight,
 			],
 			'output'      => [
 				[