+++ lib/init.php
@@ -238,6 +238,7 @@ function mai_load_files() {
 
 		// Functions.
 		'functions/autoload',
+		'functions/colors',
 		'functions/defaults',
 		'functions/deprecated',
 		'functions/enqueue',
@@ -285,6 +286,7 @@ function mai_load_files() {
 		'blocks/heading',
 		'blocks/icon',
 		'blocks/image',
+		'blocks/settings',
 		'blocks/paragraph',
 		'blocks/social-links',